Use Netlify build script and merge generators
