--- .github/workflows/generate_tools.yml
@@ -1,34 +1,0 @@
-name: Generate site data
-
-on:
-  push:
-    branches: [ main, work ]
-  workflow_dispatch:
-
-jobs:
-  build:
-    runs-on: ubuntu-latest
-    permissions:
-      contents: write
-    steps:
-      - uses: actions/checkout@v3
-        with:
-          token: ${{ secrets.GITHUB_TOKEN }}
-      - uses: actions/setup-node@v3
-        with:
-          node-version: 22
-      - name: Generate tools.json
-        run: node generate-tools.js
-      - name: Generate external-sites.json
-        run: node generate-external-sites.js
-      - name: Commit generated files
-        run: |
-          git config user.name github-actions
-          git config user.email github-actions@github.com
-          git add tools.json external-sites.json
-          if git diff --cached --quiet; then
-            echo "No changes"
-          else
-            git commit -m "Update generated data"
-            git push
-          fi
