+++ netlify.toml
@@ -1,0 +1,3 @@
+[build]
+  command = "node generate.js"
+  publish = "."
